refactor and split up stats view functions
